refactor: merge modules into single file
