+++ README.md
@@ -6,6 +6,8 @@ Work in progress – removes excessive ellipsis.
 Extension สำหรับจัดการ "..." ที่โมเดลมักใช้มากเกินไป
 โดยจะลบ `..`, `...` หรือ `…` ตามการตั้งค่าที่กำหนด (ค่าเริ่มต้น = ลบทั้งหมด)
 
+ปัจจุบันรวมโค้ดทั้งหมดไว้ในไฟล์ `index.js` เพียงไฟล์เดียว และมีปุ่ม `Check …` สำหรับตรวจสอบจำนวนจุดไข่ปลาที่เหลืออยู่ในบทสนทนาโดยไม่แก้ไขข้อความ
+
 ### วิธีติดตั้ง
 1. เปิด SillyTavern → Extensions → Install from URL  
 2. วางลิงก์ GitHub Repo นี้
